+++ src/pybandee/penta/numba/_ptrans1.py
@@ -16,7 +16,7 @@ from typing import Tuple
 import numpy as np
 from numpy.typing import NDArray
 
-from .._utils import jit
+from ..._utils import jit
 
 # === Functions ===
